Tweak and rename GuiEditMenuParentCount
